fix: renamed folders to match develop branch
