+++ src/test/java/driver/DriverFactory.java
@@ -1,4 +1,4 @@
-package drivers;
+package driver;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
