+++ src/test/java/pages/UserBrowserPO.java
@@ -1,4 +1,4 @@
-package page;
+package pages;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindAll;
