+++ src/test/java/stepDef/AddUserToGroupSteps.java
@@ -1,15 +1,15 @@
-package steps;
+package stepDef;
 
-import drivers.DriverFactory;
+import driver.DriverFactory;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
-import page.BulkEditGroupPO;
-import page.GroupBrowserPO;
-import page.LoginPO;
+import pages.BulkEditGroupPO;
+import pages.GroupBrowserPO;
+import pages.LoginPO;
 
 import java.util.List;
